Commit Barra Vida Enemigos (Diiego S)

+++ Stella's_Umbra/Assets/Diego.B/Scripts/AtaqueDesdeArriba.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class AtaqueDesdeArriba : MonoBehaviour
 {
-    [SerializeField] private int _damage;
+    [SerializeField] private float _damage;
 
     private void OnTriggerEnter(Collider other)
     {
